{"schema":"decodex/commit/1","summary":"Split knowledge source DTOs into a child module after strict validation.","authority":"manual"}
